refact: move order types to own file
